+++ packages/calendar/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.4](https://github.com/ing-bank/lion/compare/@lion/calendar@0.9.3...@lion/calendar@0.9.4) (2020-06-18)
+
+**Note:** Version bump only for package @lion/calendar
+
+
+
+
+
 ## [0.9.3](https://github.com/ing-bank/lion/compare/@lion/calendar@0.9.2...@lion/calendar@0.9.3) (2020-06-08)
 
 **Note:** Version bump only for package @lion/calendar