+++ packages/checkbox-group/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.10.6](https://github.com/ing-bank/lion/compare/@lion/checkbox-group@0.10.5...@lion/checkbox-group@0.10.6) (2020-06-18)
+
+**Note:** Version bump only for package @lion/checkbox-group
+
+
+
+
+
 ## [0.10.5](https://github.com/ing-bank/lion/compare/@lion/checkbox-group@0.10.4...@lion/checkbox-group@0.10.5) (2020-06-10)
 
 **Note:** Version bump only for package @lion/checkbox-group