+++ packages/steps/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.3](https://github.com/ing-bank/lion/compare/@lion/steps@0.5.2...@lion/steps@0.5.3) (2020-06-18)
+
+**Note:** Version bump only for package @lion/steps
+
+
+
+
+
 ## [0.5.2](https://github.com/ing-bank/lion/compare/@lion/steps@0.5.1...@lion/steps@0.5.2) (2020-06-08)
 
 **Note:** Version bump only for package @lion/steps