+++ packages/tabs/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.4](https://github.com/ing-bank/lion/compare/@lion/tabs@0.4.3...@lion/tabs@0.4.4) (2020-06-18)
+
+**Note:** Version bump only for package @lion/tabs
+
+
+
+
+
 ## [0.4.3](https://github.com/ing-bank/lion/compare/@lion/tabs@0.4.2...@lion/tabs@0.4.3) (2020-06-10)
 
 **Note:** Version bump only for package @lion/tabs